accounts/usbwallet: Get usbwallet to compile on FreeBSD
This should also improve a lot of issues with compilation for macos.

FreeBSD provides its own version of libusb, which is compatible with an
older version of Linux' libusb. As a result, getting the device list
doesn't automatically provide the device descriptor, which means that
calling libusb_get_device_descriptor is needed. This is still compatible
with the newer versions of libusb on Linux, so it's fine.

Also, note that FreeBSD uses the Linux version of hid.c, not the macos
one.